feat(nat): rename compute submodule as setup
