+++ nat/src/stateless/setup/collapse.rs
