+++ nat/src/stateless/setup/prefixtrie.rs
